Convert radio input into functional component
